rename: 스타일 폴더 및 파일 분리
