+++ src/components/test/loading/Loading.styles.tsx
@@ -1,0 +1,24 @@
+import tw from "twin.macro";
+
+export const LoadingSection = tw.div`
+w-[390px]
+h-[790.96px] 
+flex
+flex-col	
+items-center
+justify-center	
+absolute
+bg-black
+`;
+
+export const TextTop = tw.p`
+text-white
+py-3
+text-xl	
+`;
+
+export const TextBottom = tw.p`
+text-white
+py-3
+text-lg
+`;
