+++ src/pages/test/test/Test.styles.tsx
@@ -1,0 +1,14 @@
+import tw from "twin.macro";
+
+export const TestPage = tw.section`
+flex 
+flex-col 
+items-center
+justify-center
+w-[390px] 
+h-[790.96px] 
+mx-auto 
+my-0 
+bg-black
+overflow-hidden	
+`;
